refactor: extract generateDisplayName
